add option to autodelete commands

+++ src/Command.js
@@ -20,8 +20,6 @@ module.exports = class Command {
 			aliases.map(async alias => {
 				if(config[guild_id].prefix + alias === root.toLowerCase() && config[guild_id].commands[key].enabled) {
 
-					//channel.startTyping();
-
 					await this.onCommand({
 						root, args, channel, message, guild,
 						sender: member,
@@ -29,9 +27,7 @@ module.exports = class Command {
 						audit: config[guild.id].audit.enabled ? guild.channels.cache.get(config[guild.id].audit.channel) : false
 					});
 
-					await message.delete();
-
-					//channel.stopTyping();
+					if(config[guild_id]["autoremove-successful-commands"]) await message.delete();
 
 				}
 			})
